requirements.txt
lxml==5.3.0
Pillow==10.4.0
PyExifTool==0.5.6
Requests==2.32.3
